XXX c1541: Pipeline track load.
Start writing back & loading new track data relatively to read head
position.

XXX: It turns out to not be the cause of track skew, and if this feature
turns out to be a bad idea reading and writing need to be inhibited during
this period (which was probably lost in another of my changes).